Update USART and  BLE communication (#52)
* chore(core): add missing i18n translations

* fix: update layout and fix bugs

* Touch: slide into bootloader

* Touch: return OneKey version and BLE name

* Touch: update USART communication protocol

* Touch: temper remove BLE name in get-feature

* Touch: fix device found issue with web

* Touch: update firmware version

* Touch: use string type for OneKey version

* Touch: use short type for tag in protobuf

* Touch: update USART communication protocol

* Touch: display OneKey version

* Touch: fix BLE UART communication issue

* Touch: fix protobuf short tag decode issue

* Touch: fix pystyle issue

* Touch: add missing OneKey protobuf type

* Touch: fix protobuf decode issue

+++ core/embed/rust/src/protobuf/defs.rs
@@ -25,14 +25,14 @@ impl MsgDef {
         })
     }
 
-    pub fn field(&self, tag: u8) -> Option<&FieldDef> {
+    pub fn field(&self, tag: u16) -> Option<&FieldDef> {
         self.fields.iter().find(|field| field.tag == tag)
     }
 }
 
 #[repr(C, packed)]
 pub struct FieldDef {
-    pub tag: u8,
+    pub tag: u16,
     flags_and_type: u8,
     enum_or_msg_offset: u16,
     pub name: u16,
